+++ packages/state-redis/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @chat-adapter/state-redis
 
+## 4.26.0
+
+### Patch Changes
+
+- Updated dependencies [2235c16]
+- Updated dependencies [ddb084b]
+  - chat@4.26.0
+
+## 4.25.0
+
+### Patch Changes
+
+- Updated dependencies [2700ce8]
+  - chat@4.25.0
+
 ## 4.24.0
 
 ### Patch Changes